[MileStone] CAT v2 temporarily rm all chunk-related codes merge hdf5 and  pickle convert script currently only finish recipes wsj and swbd rm python2 support
