+++ egs/thchs30/run.sh
@@ -1,3 +1,6 @@
+echo "WARNING: Not implement in CAT v2 yet. Try CAT v1 or help with the implementation."                                                                        
+exit 0
+
 . ./cmd.sh
 . ./path.sh
