+++ scripts/ctc-crf/utils.py
@@ -1,23 +1,56 @@
-import scheduler
-import sys
-import argparse
-import torch
-import shutil
-import math
-import random
-import timeit
+"""
+Copyright 2021 Tsinghua University
+Apache 2.0.
+Author: Zheng Huahuan (zhh20@mails.tsinghua.edu.cn)
+"""
+
 import os
+import time
+import json
+import shutil
+import scheduler
+from collections import OrderedDict
+from monitor import plot_monitor
+from ctc_crf import CTC_CRF_LOSS as CRFLoss
+from ctc_crf import WARP_CTC_LOSS as CTCLoss
 
 
 import torch
 import torch.distributed as dist
-
-base_line_batch_size = 256
+from torch.nn.utils.rnn import pad_sequence
 
 
 def GetScheduler(scheduler_configs, param_list) -> scheduler.Scheduler:
     schdl_base = getattr(scheduler, scheduler_configs['type'])
     return schdl_base(scheduler_configs['optimizer'], param_list, **scheduler_configs['kwargs'])
+
+
+def pad_list(xs, pad_value=0, dim=0):
+    """Perform padding for the list of tensors.
+
+    Args:
+        xs (`list`): List of Tensors [(T_1, `*`), (T_2, `*`), ..., (T_B, `*`)].
+        pad_value (float): Value for padding.
+
+    Returns:
+        Tensor: Padded tensor (B, Tmax, `*`).
+
+    Examples:
+        >>> x = [torch.ones(4), torch.ones(2), torch.ones(1)]
+        >>> x
+        [tensor([1., 1., 1., 1.]), tensor([1., 1.]), tensor([1.])]
+        >>> pad_list(x, 0)
+        tensor([[1., 1., 1., 1.],
+                [1., 1., 0., 0.],
+                [1., 0., 0., 0.]])
+
+    """
+    if dim == 0:
+        return pad_sequence(xs, batch_first=True, padding_value=pad_value)
+    else:
+        xs = [x.transpose(0, dim) for x in xs]
+        padded = pad_sequence(xs, batch_first=True, padding_value=pad_value)
+        return padded.transpose(1, dim+1).contiguous()
 
 
 def str2num(src: str) -> list:
@@ -141,276 +174,259 @@ class AverageMeter(object):
 
 
 def highlight_msg(msg: str):
-    len_str = len(msg)
+    terminal_col = os.get_terminal_size().columns
+    max_len = terminal_col-4
+    if max_len <= 0:
+        print(msg)
+        return None
+
     if '\n' in msg:
         msg = msg.split('\n')
-        len_str = max([len(line) for line in msg])
-        for i, line in enumerate(msg):
-            left_pad = (len_str-len(line))//2
-            right_pad = len_str-len(line) - left_pad
-            msg[i] = '# ' + left_pad*' ' + line + right_pad*' ' + ' #'
-        msg = '\n'.join(msg)
+        len_msg = max([len(line) for line in msg])
     else:
-        msg = '# ' + msg + ' #'
-    print('\n' + "#"*(len_str + 4))
+        len_msg = len(msg)
+        msg = [msg]
+
+    if len_msg > max_len:
+        len_msg = max_len
+        new_msg = []
+        for line in msg:
+            if len(line) > max_len:
+                _cur_msg = [line[i*max_len:(i+1)*max_len]
+                            for i in range(len(line)//max_len+1)]
+                new_msg += _cur_msg
+            else:
+                new_msg.append(line)
+        del msg
+        msg = new_msg
+
+    for i, line in enumerate(msg):
+        right_pad = len_msg-len(line)
+        msg[i] = '# ' + line + right_pad*' ' + ' #'
+    msg = '\n'.join(msg)
+
+    msg = '\n' + "#"*(len_msg + 4) + '\n' + msg
+    msg += '\n' + "#"*(len_msg + 4) + '\n'
     print(msg)
-    print("#"*(len_str + 4) + '\n')
 
 
-def save_ckpt(state, is_best, ckpt_path, filename):
-    torch.save(state, '{}/{}'.format(ckpt_path, filename))
-    if is_best:
-        shutil.copyfile('{}/{}'.format(ckpt_path, filename),
-                        '{}/best_model'.format(ckpt_path))
+class Manager(object):
+    def __init__(self, logger: OrderedDict, func_buil_model, args):
+        super().__init__()
 
+        with open(args.config, 'r') as fi:
+            configures = json.load(fi)
 
-def train(model, tr_dataloader, optimizer, epoch, args, logger):
-    model.train()
-    prev_t = timeit.default_timer()
-    for i, minibatch in enumerate(tr_dataloader):
-        try:
-            features, input_lengths, labels_padded, label_lengths, path_weights = minibatch
-            if (features.size(0) < args.gpu_batch_size):
+        self.model = func_buil_model(args, configures)
+
+        self.scheduler = GetScheduler(
+            configures['scheduler'], self.model.parameters())
+
+        self.log = logger
+        self.rank = args.rank
+        self.DEBUG = args.debug
+
+        if args.resume is not None:
+            print(f"Resuming from: {args.resume}")
+            loc = f'cuda:{args.gpu}'
+            checkpoint = torch.load(args.resume, map_location=loc)
+            self.load(checkpoint)
+
+    def run(self, train_sampler, trainloader, testloader, args):
+
+        epoch = self.scheduler.epoch_cur
+        self.model.train()
+        while True:
+            epoch += 1
+            train_sampler.set_epoch(epoch)
+
+            train(trainloader, epoch, args, self)
+
+            self.model.eval()
+            metrics = test(testloader, args, self)
+            if isinstance(metrics, tuple):
+                # defaultly use the first one to evaluate
+                metrics = metrics[0]
+            state = self.scheduler.step(epoch, metrics)
+
+            self.model.train()
+            if self.rank == 0 and not self.DEBUG:
+                self.log_export(args.ckptpath)
+                plot_monitor(args.dir.split('/')[-1])
+
+            if state == 2:
+                print("Break: GPU[%d]" % self.rank)
+                dist.barrier()
+                break
+            elif self.rank != 0 or self.DEBUG:
                 continue
-            model.zero_grad()
-            optimizer.zero_grad()
+            elif state == 0 or state == 1:
+                self.save("checkpoint", args.ckptpath)
+                if state == 1:
+                    shutil.copyfile(
+                        f"{args.ckptpath}/checkpoint.pt", f"{args.ckptpath}/bestckpt.pt")
 
-            loss = model(features, labels_padded, input_lengths, label_lengths)
-            if args.ctc_crf:
-                partial_loss = torch.mean(loss.cpu())
+                    # save model for inference
+                    torch.save(self.model.module.infer.state_dict(),
+                               args.ckptpath + "/infer.pt")
+            else:
+                raise ValueError(f"Unknown state: {state}.")
+            torch.cuda.empty_cache()
+
+    def save(self, name, PATH=''):
+        """Save checkpoint.
+
+        The checkpoint file would be located at `PATH/name.pt`
+        or `name.pt` if `PATH` is empty.
+        """
+
+        if PATH != '' and PATH[-1] != '/':
+            PATH += '/'
+        torch.save(OrderedDict({
+            'model': self.model.state_dict(),
+            'scheduler': self.scheduler.state_dict(),
+            'log': OrderedDict(self.log)
+        }), PATH+name+'.pt')
+
+    def load(self, checkpoint):
+        r'Load checkpoint.'
+
+        dist.barrier()
+        self.model.load_state_dict(checkpoint['model'])
+        self.scheduler.load_state_dict(checkpoint['scheduler'])
+        self.log = checkpoint['log']
+
+    def log_update(self, msg=[], loc="log_train"):
+        assert loc in self.log
+
+        self.log[loc].append(msg)
+
+    def log_export(self, PATH):
+        """Save log file in {PATH}/{key}.csv
+        """
+
+        for key, value in self.log.items():
+
+            with open(f"{PATH}/{key}.csv", 'w+', encoding='utf8') as file:
+                data = [','.join([str(x) for x in infos])
+                        for infos in value[1:]]
+                file.write(value[0] + '\n' + '\n'.join(data))
+
+
+def train(trainloader, epoch: int, args, manager: Manager):
+
+    scheduler = manager.scheduler
+
+    model = manager.model
+    optimizer = scheduler.optimizer
+
+    batch_time = AverageMeter('Time', ':6.3f')
+    data_time = AverageMeter('Data', ':6.3f')
+    losses = AverageMeter('Loss', ':.4e')
+    losses_real = AverageMeter('Loss_real', ':.4e')
+    progress = ProgressMeter(
+        len(trainloader),
+        [batch_time, data_time, losses, losses_real],
+        prefix="Epoch: [{}]".format(epoch))
+
+    end = time.time()
+
+    for i, minibatch in enumerate(trainloader):
+        if args.debug and i > 20:
+            highlight_msg("In debug mode, quit training.")
+            dist.barrier()
+            break
+        # measure data loading time
+        logits, input_lengths, labels, label_lengths, path_weights = minibatch
+
+        data_time.update(time.time() - end)
+
+        optimizer.zero_grad()
+
+        logits, labels, input_lengths, label_lengths = logits.cuda(
+            args.gpu, non_blocking=True), labels, input_lengths, label_lengths
+
+        loss = model(logits, labels, input_lengths, label_lengths)
+
+        with torch.no_grad():
+            if args.iscrf:
+                partial_loss = loss.cpu()
                 weight = torch.mean(path_weights)
                 real_loss = partial_loss - weight
             else:
-                real_loss = torch.mean(loss.cpu())
-            loss.backward()
-            optimizer.step()
+                real_loss = loss.cpu()
 
-            t2 = timeit.default_timer()
-            if i % 200 == 0 and args.rank == 0:
-                logger.debug("epoch: {}, step: {}, time: {}, tr_real_loss: {}, lr: {}".format(
-                    epoch, i, t2 - prev_t, real_loss.item(), optimizer.param_groups[0]['lr']))
-            prev_t = t2
-        except Exception as ex:
-            print("rank {} train exception ".format(args.rank), ex)
+        # measure accuracy and record loss
+        losses.update(loss.item(), logits.size(0))
+        losses_real.update(real_loss.item(), logits.size(0))
+
+        loss.backward()
+
+        optimizer.step()
+        scheduler.update_lr((epoch - 1) * len(trainloader) + i + 1)
+
+        # measure elapsed time
+        batch_time.update(time.time() - end)
+        manager.log_update(
+            [epoch, loss.item(), real_loss.item(), optimizer.param_groups[0]['lr'], time.time() - end], loc='log_train')
+
+        end = time.time()
+
+        if i % args.print_freq == 0 or args.debug:
+            progress.display(i)
 
 
-def validate(model, cv_dataloader, epoch, args, logger):
-    # cv stage
-    model.eval()
-    cv_total_loss = 0
-    cv_total_sample = 0
+def test(testloader, args, manager: Manager):
+
+    model = manager.model
+
+    batch_time = AverageMeter('Time', ':6.3f')
+    data_time = AverageMeter('Data', ':6.3f')
+    losses_real = AverageMeter('Loss_real', ':.4e')
+    progress = ProgressMeter(
+        len(testloader),
+        [batch_time, data_time, losses_real],
+        prefix='Test: ')
+
+    beg = time.time()
+    end = time.time()
     with torch.no_grad():
-        for i, minibatch in enumerate(cv_dataloader):
-            try:
-                features, input_lengths, labels_padded, label_lengths, path_weights = minibatch
-                loss = model(features, labels_padded,
-                             input_lengths, label_lengths)
-                if args.ctc_crf:
-                    partial_loss = torch.mean(loss.cpu())
-                    weight = torch.mean(path_weights)
-                    real_loss = partial_loss - weight
-                else:
-                    real_loss = torch.mean(loss.cpu())
-                cv_total_loss += real_loss.item() * features.size(0)
-                cv_total_sample += features.size(0)
-            except Exception as ex:
-                print("rank {} cv exception ".format(args.rank), ex)
-        cv_loss = cv_total_loss / cv_total_sample
-        if args.rank == 0:
-            logger.info("epoch: {}, mean_cv_loss: {}".format(epoch, cv_loss))
-    return cv_loss
+        for i, minibatch in enumerate(testloader):
+            if args.debug and i > 20:
+                highlight_msg("In debug mode, quit evaluating.")
+                dist.barrier()
+                break
+            # measure data loading time
+            logits, input_lengths, labels, label_lengths, path_weights = minibatch
+            data_time.update(time.time() - end)
 
+            logits, labels, input_lengths, label_lengths = logits.cuda(
+                args.gpu, non_blocking=True), labels, input_lengths, label_lengths
+            path_weights = path_weights.cuda(args.gpu, non_blocking=True)
 
-def train_chunk_model(model, reg_model, tr_dataloader, optimizer, epoch, chunk_size, TARGET_GPUS, args, logger):
-    prev_t = 0
-    for i, minibatch in enumerate(tr_dataloader):
-        try:
-            logits, input_lengths, labels_padded, label_lengths, path_weights = minibatch
-            model.zero_grad()
-            optimizer.zero_grad()
-            input_lengths = map(lambda x: x.size()[0], logits)
-            if sys.version > '3':
-                input_lengths = list(input_lengths)
+            loss = model(logits, labels, input_lengths, label_lengths)
 
-            input_lengths = torch.IntTensor(input_lengths)
-            out1_reg, out2_reg, out3_reg = reg_model(
-                logits, labels_padded, input_lengths, label_lengths)
-            loss, loss_cls, loss_reg = model(logits, labels_padded, input_lengths, label_lengths,
-                                             chunk_size, out1_reg.detach(), out2_reg.detach(), out3_reg.detach())
+            if args.iscrf:
+                weight = torch.mean(path_weights)
+                real_loss = loss - weight
+            else:
+                real_loss = loss
 
-            partial_loss = torch.mean(loss.cpu())
-            loss_cls = torch.mean(loss_cls.cpu())
-            loss_reg = torch.mean(loss_reg.cpu())
+            dist.all_reduce(real_loss, dist.ReduceOp.SUM)
+            real_loss = real_loss / torch.cuda.device_count()
 
-            weight = torch.mean(path_weights)
-            real_loss = partial_loss - weight
-            loss_cls = loss_cls - weight
+            # measure accuracy and record loss
+            losses_real.update(real_loss.item(), logits.size(0))
 
-            count = min(loss.size(0), len(TARGET_GPUS))
-            loss.backward(loss.new_ones(count))
+            # measure elapsed time
+            batch_time.update(time.time() - end)
 
-            optimizer.step()
-            t2 = timeit.default_timer()
-            if i % 200 == 0 and args.rank == 0:
-                logger.debug("rank {} epoch:{} step:{} time: {}, tr_real_loss: {},loss_cls: {},loss_reg:{}, lr: {}".format(
-                    args.rank, epoch, i, t2 - prev_t, real_loss.item(), loss_cls.item(), loss_reg.item(), optimizer.param_groups[0]['lr']))
-            prev_t = t2
-            torch.cuda.empty_cache()
-        except Exception as ex:
-            print("rank {} train exception ".format(args.rank), ex)
+            end = time.time()
 
+            if i % args.print_freq == 0 or args.debug:
+                progress.display(i)
 
-def validate_chunk_model(model, reg_model, cv_dataloader, epoch, cv_losses_sum, cv_cls_losses_sum, args, logger):
-    count = 0
-    for i, minibatch in enumerate(cv_dataloader):
-        try:
-            #print("cv epoch: {}, step: {}".format(epoch, i))
-            logits, input_lengths, labels_padded, label_lengths, path_weights = minibatch
+    manager.log_update(
+        [losses_real.avg, time.time() - beg], loc='log_eval')
 
-            input_lengths = map(lambda x: x.size()[0], logits)
-            if sys.version > '3':
-                input_lengths = list(input_lengths)
-            input_lengths = torch.IntTensor(input_lengths)
-
-            reg_out1, reg_out2, reg_out3 = reg_model(
-                logits, labels_padded, input_lengths, label_lengths)
-            loss, loss_cls, loss_reg = model(logits, labels_padded, input_lengths, label_lengths,
-                                             args.default_chunk_size, reg_out1.detach(), reg_out2.detach(), reg_out3.detach())
-
-            loss_size = loss.size(0)
-            count += loss_size
-
-            partial_loss = torch.mean(loss.cpu())
-            weight = torch.mean(path_weights)
-            real_loss = partial_loss - weight
-
-            loss_cls = torch.mean(loss_cls.cpu())
-            loss_cls = loss_cls - weight
-            loss_reg = torch.mean(loss_reg.cpu())
-
-            real_loss_sum = real_loss * loss_size
-            loss_cls_sum = loss_cls * loss_size
-
-            cv_losses_sum.append(real_loss_sum.item())
-            cv_cls_losses_sum.append(loss_cls_sum.item())
-        except Exception as ex:
-            print("rank {} cv exception ".format(args.rank), ex)
-    return count
-
-
-def adjust_lr(optimizer, origin_lr, lr, cv_loss, prev_cv_loss, epoch, min_epoch):
-    if epoch < min_epoch or cv_loss <= prev_cv_loss:
-        pass
-    else:
-        lr = lr / 10.0
-    for param_group in optimizer.param_groups:
-        param_group['lr'] = lr
-
-
-def adjust_lr_distribute(optimizer, origin_lr, lr, cv_loss, prev_cv_loss, epoch, annealing_epoch, gpu_batch_size, world_size):
-    '''
-    The hyperparameter setup for the batch size 256
-    configuration is the learning rate is set to be 0.1, momentum is
-    set as 0.9, and learning rate anneals by p12 every epoch from the
-    11th epoch. The training finishes in 16 epochs. Inspired by the
-    work proposed in[4], we are able to increase the batch size from
-    256 to 2560 without decreasing model accuracy by (1) linearly
-    warming up the base learning rate from 0.1 to 1 in the first 10
-    epochs and (2) annealing the learning rate by p12 from the 11th
-    Epoch. refer Wei Zhang, Xiaodong Cui, Ulrich Finkler, Brian Kingsbury,
-    George Saon, David Kung, Michael Picheny "Distributed Deep Learning 
-    Strategies For Automatic Speech Recognition"
-    '''
-
-    new_lr = lr
-    batch_size = gpu_batch_size * world_size
-    if epoch < annealing_epoch:
-        if batch_size > base_line_batch_size:
-            max_lr = (batch_size/base_line_batch_size) * origin_lr
-            new_lr = lr + round(max_lr/annealing_epoch, 6)
-        else:
-            new_lr = origin_lr
-    else:
-        new_lr = lr / math.sqrt(2)
-
-    for param_group in optimizer.param_groups:
-        param_group['lr'] = new_lr
-
-    return new_lr
-
-
-def parse_args():
-    parser = argparse.ArgumentParser(description="recognition argument")
-    parser.add_argument("--model", type=str,
-                        help="the model name will be trained")
-    parser.add_argument(
-        "--arch",
-        choices=[
-            'BLSTM', 'LSTM', 'VGGBLSTM', 'VGGLSTM', 'LSTMrowCONV', 'TDNN_LSTM',
-            'BLSTMN'
-        ],
-        default='BLSTM')
-    parser.add_argument("--ctc_crf", action="store_true", default=False,
-                        help="whether to use ctc_crf or not, true for ctc_crf, false for ctc")
-    parser.add_argument("--min_epoch", type=int, default=10)
-    parser.add_argument("--lamb", type=float, default=0.1)
-    parser.add_argument("--hdim", type=int, default=512)
-    parser.add_argument("--layers", type=int, default=6)
-    parser.add_argument("--output_unit", type=int,
-                        help='the phone number of your model output, must be specified')
-    parser.add_argument("--dropout", type=float, default=0.5)
-    parser.add_argument("--feature_size", type=int, default=120)
-    parser.add_argument("--origin_lr", type=float, default=0.001)
-    parser.add_argument("--annealing_epoch", type=int, default=15)
-    parser.add_argument("--data_loader_workers", type=int, default=16)
-    parser.add_argument("--reg_weight", type=float, default=0.01)
-    parser.add_argument("--stop_lr", type=float, default=0.00001)
-    parser.add_argument("--batch_size",  type=int, default=32,
-                        help='batch size for non distribute training')
-    parser.add_argument("--tr_data_path", type=str, help="training data path")
-    parser.add_argument("--dev_data_path", type=str,
-                        help="validation data path")
-    parser.add_argument("--den_lm_fst_path", type=str, default=None,
-                        help="denominator fst path, be used in ctc_crf")
-    parser.add_argument("--checkpoint", type=str, default=None,
-                        help="the checkpoint file path of pretrained model")
-    parser.add_argument('--dist_url', type=str, default=None,
-                        help='Url used to set up distributed training. It must be the IP of device running the rank 0, \
-                        and a free port can be opened and listened by pytorch distribute framework.')
-    parser.add_argument('--world_size', type=int, default=0, help='world size must be set to indicate the number of ranks you \
-                        will  start to train your model. If the real rank number is less then the value of \
-                        --world_size it will be blocked on function init_process_group. Rather than that if a rank \
-                        number grater then the value of world_size was given the function init_process_group will \
-                        throw a exception to terminate the training. So you need make sure the world size is equal \
-                        to the ranks number.')
-    parser.add_argument("--start_rank", type=int, default=0, help='This value was used to specify the start rank on the device. \
-                        For example, if you have 3 gpu and 3 ranks on device 1 and the 4 gpu and 4 ranks on \
-                        device 2. The device 1 has rank 0, then the device 2 must start from rank 3.So you must \
-                        set --start_rank 3 on device 2. ')
-
-    parser.add_argument('--gpu_batch_size', default=32, type=int,
-                        help='This value was depended on the memory size of your gpu, the biger the value was set, \
-                        the higher training speed you can get. The total batch size of your training was \
-                        (world size) * (batch size per gpu). The max batch size must not grater then 2560, \
-                        or rather the training will not be convergent.')
-
-    parser.add_argument("--spec_augment", action="store_true", default=False,
-                        help="whether to use specAugment in training phase or not, true for using")
-
-    parser.add_argument("--regmodel_checkpoint", type=str, default=None,
-                        help='the checkpoint file path of reg model')
-
-    parser.add_argument("--default_chunk_size", type=int, default=40,
-                        help='default chunk size for low latency model')
-
-    parser.add_argument("--jitter_range", type=int, default=10,
-                        help='contextual frame number of low latency model')
-    parser.add_argument("--cate", type=int, default=4000,
-                        help='the number of pkls the training ark data was diviced and assigned to by utterence length')
-
-    args = parser.parse_args()
-
-    args.csv_file = args.model + ".csv"
-    args.figure_file = args.model + ".png"
-
-    return args
+    return losses_real.avg
